Add `hostname` command; Fix TCP server termination.
Hostname setting implementation: instead of the "LWIP", set a default hostname in the format "ESP-XXXXXX" where XXXXXX are the last 3 octets of the board's base MACAddr in hexadecimal format;
Also, fixed broken source-code formatting (indenting/spacing) in event_handler() and wifi_cmd_ap_set().

Rearranged register_autorun() to group code/definitions by command being registered, much clearer that way.

Removed duplicate #define IPERF_REPORT_TASK_NAME

Added `hostname` console command
Reformatted text of help for the `autorun_set` command for better legibility;
Rearranged and commented the code and definitions for registering the console commands for better legibility.

Make iperf_traffic_task() finish when the client does;
Rearranged iperf_traffic_task() so as to eliminate unneeded while() loop and overall make it clearer;
Instead of only terminating the recv() loop when it returns <0, now also terminates when it returns =0 (meaning the connection was reset because the client finished and closed it).

Added notes on new `hostname` command, Etc;

Etc1: fix for the TCP server mode, to no longer wait until test time exhausted to finish but rather to finish along with the client.
Etc2: the client mode was now tested (not as deeply as the server mode) and found to work.